requirements.txt
# OFI论文实验项目依赖
# 安装: pip install -r requirements.txt

# ============================================================
# 数据处理与分析
# ============================================================
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# ============================================================
# 数据库
# ============================================================
psycopg2-binary>=2.9.9      # PostgreSQL连接
python-dotenv>=1.0.0        # 环境变量加载

# ============================================================
# 数据存储格式
# ============================================================
pyarrow>=14.0.0             # Parquet文件支持

# ============================================================
# 富途API
# ============================================================
futu-api>=8.10.0            # 富途OpenAPI SDK

# ============================================================
# 可视化（可选）
# ============================================================
matplotlib>=3.7.0
seaborn>=0.12.0

# ============================================================
# 机器学习与深度学习
# ============================================================
scikit-learn>=1.3.0         # 机器学习基准模型
xgboost>=2.0.0              # XGBoost模型
torch>=2.0.0                # PyTorch深度学习框架

# ============================================================
# 模型解释（后续使用）
# ============================================================
# shap>=0.42.0              # SHAP特征归因
